+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `help` command - sort commands alphabetically
-- (Docker) `REALIP_FROM_CLOUDFLARE` environment variable for setting X-Real-IP from CF-Connecting-IP
+- (Docker) Support setting Nginx `real_ip_from`, see docs for environment variables 'REALIP_SOURCE',
+  'REALIP_CUSTOM_FROM', 'REALIP_CUSTOM_HEADER'
 - BREAKING: (Docker) Support host & bridge networking
 
 ## [2.0.0]